Изменить сигнатуру тестов, добавить дженерики.
