+++ test/managers/FileBackedTaskManagerTest.java
@@ -3,9 +3,12 @@ package managers;
 import exceptions.ManagerSaveException;
 import manager.FileBackedTaskManager;
 import manager.TaskManager;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import task.Epic;
 import task.Status;
+import task.Subtask;
 import task.Task;
 
 import java.io.File;
@@ -17,6 +20,29 @@ import java.time.LocalDateTime;
 import static org.junit.jupiter.api.Assertions.*;
 
 class FileBackedTaskManagerTest extends TaskManagerTest<FileBackedTaskManager> {
+    File file;
+
+    @BeforeEach
+    public void beforeEach() {
+        try {
+            file = File.createTempFile("temp", ".csv");
+            taskManager = new FileBackedTaskManager(file);
+        } catch (IOException ignored) {
+
+        }
+        task1 = new Task("Test 1", "Description 1", Status.IN_PROGRESS,
+                LocalDateTime.now(), Duration.ofMinutes(5));
+        task2 = new Task("Test 2", "Description 2", Status.IN_PROGRESS,
+                LocalDateTime.now().plusMinutes(10), Duration.ofMinutes(5));
+        epic1 = new Epic(1, "Test 1", "Description 1");
+        taskManager.addNewEpic(epic1);
+        int epicId = epic1.getId();
+        epic2 = new Epic(2, "Test 2", "Description 2");
+        subtask1 = new Subtask("Test 1-1", "Description 1", Status.NEW,
+                LocalDateTime.now().plusMinutes(45), Duration.ofMinutes(5), epicId);
+        subtask2 = new Subtask("Test 1-2", "Description 2", Status.DONE,
+                LocalDateTime.now().plusMinutes(60), Duration.ofMinutes(5), epicId);
+    }
 
 
     @Test
@@ -56,13 +82,4 @@ class FileBackedTaskManagerTest extends TaskManagerTest<FileBackedTaskManager> {
         }, "should result in an error");
     }
 
-    @Override
-    TaskManager initTaskManager() {
-        try {
-            File tempFile = File.createTempFile("temp", ".csv");
-            return FileBackedTaskManager.loadFromFile(tempFile);
-        } catch (IOException ignored) {
-        }
-        return null;
-    }
 }
